"fix typo in `diff --git` for `/src/faceapi/routers/models.py.py` #4796-4806.py `git diff --git-a/b2e516d`"

+++ src/faceapi/routers/models.py
@@ -37,7 +37,7 @@ class PromptResponse(BaseResponse):
     width: Optional[int] = None
     height: Optional[int] = None
     negative_prompt: Optional[str] = None
-    seed: Optional[int] = None
+    seed: Optional[float] = None
     strength: Optional[float] = None
